+++ src/base/ScrollListener.java
@@ -4,5 +4,5 @@ package base;
  * Created by galafit on 16/11/17.
  */
 public interface ScrollListener {
-    public void onScrollChanged(double scrollValue, double scrollExtent);
+    public void onScrollChanged(float scrollValue, float scrollExtent);
 }
