+++ src/base/scales/Scale.java
@@ -3,33 +3,33 @@ package base.scales;
 import base.config.LabelFormatInfo;
 
 public abstract class Scale {
-    protected double domain[] = {0, 1};
-    protected double range[] = {0, 1};
+    protected float domain[] = {0, 1};
+    protected float range[] = {0, 1};
 
-    public void setDomain(double... domain) {
+    public void setDomain(float... domain) {
         this.domain = domain;
     }
 
-    public void setRange(double... range) {
+    public void setRange(float... range) {
         this.range = range;
     }
 
 
-    public double[] getDomain() {
+    public float[] getDomain() {
         return domain;
     }
 
-    public double[] getRange() {
+    public float[] getRange() {
         return range;
     }
 
-    public abstract double scale(double value);
+    public abstract float scale(float value);
 
-    public abstract double invert(double value);
+    public abstract float invert(float value);
 
     public abstract TickProvider getTickProvider(int tickCount, LabelFormatInfo labelFormatInfo);
 
-    public abstract TickProvider getTickProvider(double tickStep, Unit tickUnit, LabelFormatInfo labelFormatInfo);
+    public abstract TickProvider getTickProvider(float tickStep, Unit tickUnit, LabelFormatInfo labelFormatInfo);
 
 }
