+++ src/data/RegularColumn.java
@@ -7,15 +7,15 @@ import data.series.IntSeries;
  * Created by galafit on 1/11/17.
  */
 public class RegularColumn implements NumberColumn{
-    private double startValue;
-    private double dataInterval;
+    private float startValue;
+    private float dataInterval;
 
-    public RegularColumn(double startValue, double dataInterval) {
+    public RegularColumn(float startValue, float dataInterval) {
         this.startValue = startValue;
         this.dataInterval = dataInterval;
     }
 
-    public double getDataInterval() {
+    public float getDataInterval() {
         return dataInterval;
     }
 
@@ -29,19 +29,19 @@ public class RegularColumn implements NumberColumn{
     }
 
     @Override
-    public double getValue(int index) {
+    public float getValue(int index) {
         return startValue + dataInterval * index;
     }
 
     @Override
     public Range getExtremes(int from, int length) {
-        double min = startValue;
-        double max = startValue + (size() - 1) * dataInterval;
+        float min = startValue;
+        float max = startValue + (size() - 1) * dataInterval;
         return new Range(min, max);
     }
 
     @Override
-    public int upperBound(double value, int from, int length) {
+    public int upperBound(float value, int from, int length) {
         int lowerBound = lowerBound(value, from, length);
         if(value == getValue(lowerBound)) {
             return lowerBound;
@@ -55,7 +55,7 @@ public class RegularColumn implements NumberColumn{
     }
 
     @Override
-    public int lowerBound(double value, int from, int length) {
+    public int lowerBound(float value, int from, int length) {
         return (int) ((value - startValue) / dataInterval);
     }
